+++ client/app/(auth)/_layout.tsx
@@ -1,6 +1,6 @@
 import { useAuth } from '@/providers/authProvider';
 import { Redirect, Stack } from 'expo-router';
-import { NavigationHeader } from '@/components/ui/Navigation/NavigationHeader';
+import { NavigationHeader } from '@/components/modules/Navigation/NavigationHeader';
 
 export default function AuthLayout() {
   const { session } = useAuth();
