+++ client/app/(user)/profile/_layout.tsx
@@ -1,5 +1,5 @@
 import { Stack } from 'expo-router';
-import { NavigationHeader } from '@/components/ui/Navigation/NavigationHeader';
+import { NavigationHeader } from '@/components/modules/Navigation/NavigationHeader';
 
 export default function ProfileStack() {
   return (
