+++ client/components/ui/index.ts
@@ -5,4 +5,4 @@ export * from './Button/Button';
 export * from './Input/Input';
 export * from './Card/Card';
 export * from './Layout/Layout';
-export * from './Navigation';
+export * from '../modules/Navigation';
